leave FormField as json object
This is part of the way there towards matching the behavior of the iOS SDK where form fields are just left as dictionaries. We ran into an issue where form fields weren't accurately represented in the javascript on Android. This is because the FormField POJO in the Android SDK wasn't up to date.

Down the road, we'd like to change this to a HashMap (instead of JsonNode), but that will require upgrading React-Native to 0.46